+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - Add monks martial arts die as die advancement. @elwin1410
+- 2024 magic items will now use the 2024 DMG prices.
 
 # 6.3.3