+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next Up
+# 6.3.4
 
 - Add monks martial arts die as die advancement. @elwin1410
 - 2024 magic items will now use the 2024 DMG prices.